refactor(resource): split fanout orchestrator, extract runtime teardown/acquire_loop, colocate revoke types
Behavior-preserving module reorg in nebula-resource (private behind sdk, 0 external semver):
- credential_fanout: dispatch* orchestrator -> orchestrator.rs; index.rs is now a pure data structure
- runtime: teardown_deadline/destroy_within -> teardown.rs; framework acquire loop -> acquire_loop.rs (fixes resident->managed edge)
- manager: TaintedSlot/RevokeTail -> rotation.rs (sole consumer)

clippy -D warnings clean; tests green under default and --features rotation (verified this session); public paths unchanged.
